Refactor into separate classes with config
