+++ LoLClip/Usings.cs
@@ -1,0 +1,5 @@
+global using System.Net.Http;
+global using System.Net.WebSockets;
+global using System.Text;
+global using System.Text.Json;
+global using Microsoft.Extensions.Logging;
